+++ python/GafferSceneUI/__init__.py
@@ -183,6 +183,7 @@ from . import SetQueryUI
 from . import MeshSegmentsUI
 from . import ImageToPointsUI
 from . import MeshSplitUI
+from . import FramingConstraintUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.